requirements.txt
kougen-fastapi-crud
